--- src/pages/BookingPage/calendar.css
@@ -1,8 +1,0 @@
-.react-datepicker {
-  
-}
-
-.react-datepicker__header {
-  width: 600px !important;
-  height: 100px !important;
-}
